+++ com.unity.build-report-inspector/Editor/Mobile/AssembyInfo.cs
@@ -1,0 +1,4 @@
+﻿using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("Unity.BuildReportInspector.Editor")]
+[assembly: InternalsVisibleTo("Unity.BuildReportInspector.Editor.Tests.Mobile")]