+++ com.unity.build-report-inspector/Editor/Mobile/IPlatformUtilities.cs
@@ -1,0 +1,7 @@
+namespace Unity.BuildReportInspector.Mobile
+{
+    internal interface IPlatformUtilities
+    {
+        MobileArchInfo[] GetArchitectureInfo(string applicationPath);
+    }
+}