+++ com.unity.build-report-inspector/CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.2.0-preview] - 2020-04-02
+### Added
+- Add improved support for iOS, Apple TV and Android platforms.
+
+### Fixed
+- Fix #1 - incomplete Assets list shown when there are many entries.
+- Fix .NET 3.5 support.
+
 ## [0.1.2-preview] - 2019-12-18
 
 *Remove warning messages reported in https://issuetracker.unity3d.com/issues/errors-regarding-missing-meta-files-when-installing-build-report-inspector-package*